feat(sidenav): refactor package structure and some component names
affects: @crave/farmblocks-sidenav
