+++ packages/sidenav/src/components/NavButton/NavButton.js
@@ -2,21 +2,13 @@ import React from "react";
 import PropTypes from "prop-types";
 import { colors } from "@crave/farmblocks-theme";
 
-import { IconButton, CloseButton, ToggleButton } from "./NavButton.styled";
-import { PUSH, FULLSCREEN, OVERLAY } from "./constants/variants";
+import { IconButton } from "./NavButton.styled";
+import { PUSH, FULLSCREEN, OVERLAY } from "../../constants/variants";
 
 const NavButton = React.memo(({ component: Component, icon, ...props }) => (
   <Component role="button" {...props}>
     <i className={icon} />
   </Component>
-));
-
-export const NavToggle = React.memo(props => (
-  <NavButton component={ToggleButton} icon="wg-list" {...props} />
-));
-
-export const NavClose = React.memo(props => (
-  <NavButton component={CloseButton} icon="wg-close" {...props} />
 ));
 
 NavButton.defaultProps = {
