+++ packages/sidenav/src/components/NavButton/index.js
@@ -1,0 +1,3 @@
+export { default } from "./NavButton";
+export { default as CloseButton } from "./CloseButton";
+export { default as ToggleButton } from "./ToggleButton";
