+++ packages/sidenav/src/components/SideBar/index.js
@@ -1,0 +1,1 @@
+export { default, SideBarBase } from "./Sidebar";
